move the compare 'models' into views
