+++ src/lib/src/model/compare/tabular_compare_summary.rs
@@ -1,8 +1,8 @@
-use serde::{Deserialize, Serialize};
-#[derive(Deserialize, Serialize, Debug, Clone)]
-pub struct TabularCompareSummary {
-    pub num_left_only_rows: usize,
-    pub num_right_only_rows: usize,
-    pub num_diff_rows: usize,
-    pub num_match_rows: usize,
-}
+// use serde::{Deserialize, Serialize};
+// #[derive(Deserialize, Serialize, Debug, Clone)]
+// pub struct TabularCompareSummary {
+//     pub num_left_only_rows: usize,
+//     pub num_right_only_rows: usize,
+//     pub num_diff_rows: usize,
+//     pub num_match_rows: usize,
+// }
